+++ src/tool/cpp/cppwinrt/strings/base_consume.h
@@ -4,17 +4,6 @@ namespace winrt::impl
     template <typename Async>
     void blocking_suspend(Async const& async);
 
-    template <typename D> struct consume_IActivationFactory
-    {
-        template <typename T>
-        T ActivateInstance() const
-        {
-            Windows::Foundation::IInspectable instance;
-            check_hresult(WINRT_SHIM(Windows::Foundation::IActivationFactory)->ActivateInstance(put_abi(instance)));
-            return instance.try_as<T>();
-        }
-    };
-
     template <typename D, typename T> struct consume_IReference
     {
         T Value() const
