--- src/tool/cpp/cppwinrt/strings/base_shims.h
@@ -1,1 +1,0 @@
-
