--- src/tool/cpp/cppwinrt/strings/base_types.h
@@ -1,1 +1,0 @@
-
